requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
pymongo
scikit-learn
pyyaml
watchdog
dill
mlflow
fastapi
uvicorn

#-e .